requirements.txt
requests~=2.32.3
PyYAML~=6.0.1
Markdown~=3.6
huggingface-hub~=0.23.4
python-dotenv~=1.0.1
pillow~=10.4.0
diffusers~=0.29.2
pydub~=0.25.1
tiktoken~=0.7.0
Django~=5.0.6
djangorestframework~=3.15.2
pydantic~=2.8.0